Move UserMapper to Domain::MapUsersAccounts
