--- spec/lib/user_mapper_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe UserMapper do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
